+++ docs/RELEASE_NOTES.md
@@ -40,6 +40,15 @@ are breaking changes. If there are, then a migration guide should be provided.
 - [v1.0.0-M1 -  Sep 06, 2017](#v100-m1----sep-06-2017)
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
+<!--
+### v1.0.0-M5 -  Nov 7, 2017
+[Commits](https://github.com/commercetools/commercetools-sync-java/compare/v1.0.0-M4...v1.0.0-M5) |
+[Javadoc](https://commercetools.github.io/commercetools-sync-java/v/v1.0.0-M5/) | 
+[Jar](https://bintray.com/commercetools/maven/commercetools-sync-java/v1.0.0-M5)
+
+**Doc Fixes** (1)
+- **Build Tools** - Adds Snyk vulnerabilities badge to repo README. [#188](https://github.com/commercetools/commercetools-sync-java/pull/188)
+-->
 ### v1.0.0-M4 -  Nov 7, 2017
 [Commits](https://github.com/commercetools/commercetools-sync-java/compare/v1.0.0-M3...v1.0.0-M4) |
 [Javadoc](https://commercetools.github.io/commercetools-sync-java/v/v1.0.0-M4/) |